Add NoodleR roleplay authoring UI
