+++ packages/client/src/hooks/use-noodle.ts
@@ -17,6 +17,8 @@ import type {
   NoodleInteractionUpdateInput,
   NoodlePost,
   NoodlePostUpdateInput,
+  NoodlePrivatePostCreateInput,
+  NoodlePrivatePostUpdateInput,
   NoodleRemoveInteractionInput,
   NoodleRescheduleRefreshInput,
   NoodleRefreshSchedulerStatus,
@@ -25,6 +27,7 @@ import type {
   NoodleStageProfileInput,
   NoodlePrivateGenerationRequest,
   NoodleStageProfileDraftRequest,
+  NoodlerManagedPost,
   NoodlerStageProfile,
   NoodlerManagedStageProfile,
   NoodlerViewerScope,
@@ -42,11 +45,14 @@ export type NoodleRefreshResult = {
 export const noodleKeys = {
   all: ["noodle"] as const,
   bootstrap: () => [...noodleKeys.all, "bootstrap"] as const,
-  privateAccounts: () => [...noodleKeys.all, "private-accounts"] as const,
+  privateRoot: () => [...noodleKeys.all, "private"] as const,
+  privateAccounts: () => [...noodleKeys.privateRoot(), "accounts"] as const,
+  privateEligibleAccountsRoot: () => [...noodleKeys.privateRoot(), "eligible"] as const,
   privateEligibleAccounts: (search: string, kind: string) =>
-    [...noodleKeys.privateAccounts(), "eligible", search, kind] as const,
-  privatePosts: (accountId: string) => [...noodleKeys.privateAccounts(), accountId, "posts"] as const,
-  viewer: (personaId: string) => [...noodleKeys.privateAccounts(), "viewer", personaId] as const,
+    [...noodleKeys.privateEligibleAccountsRoot(), search, kind] as const,
+  privatePosts: (accountId: string) => [...noodleKeys.privateRoot(), "posts", accountId] as const,
+  privateViewers: () => [...noodleKeys.privateRoot(), "viewers"] as const,
+  viewer: (personaId: string) => [...noodleKeys.privateViewers(), personaId] as const,
 };
 
 function preservePollVotes(current: NoodleBootstrap | undefined, next: NoodleBootstrap): NoodleBootstrap {
@@ -95,7 +101,7 @@ export function useNoodlerEligibleAccounts(search: string, kind: "all" | "charac
 export function useNoodlerPosts(accountId: string | null) {
   return useQuery({
     queryKey: noodleKeys.privatePosts(accountId ?? "none"),
-    queryFn: () => api.get<NoodlePost[]>(`/noodle/noodler/accounts/${encodeURIComponent(accountId!)}/posts`),
+    queryFn: () => api.get<NoodlerManagedPost[]>(`/noodle/noodler/accounts/${encodeURIComponent(accountId!)}/posts`),
     enabled: Boolean(accountId),
     staleTime: 10_000,
   });
@@ -114,7 +120,12 @@ export function useCreateNoodlerStageProfile() {
       api.post<NoodlerStageProfile>(`/noodle/accounts/${encodeURIComponent(publicAccountId)}/private`, {
         stageProfile,
       }),
-    onSuccess: () => qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+    onSuccess: () =>
+      Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateEligibleAccountsRoot() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]),
   });
 }
 
@@ -123,7 +134,11 @@ export function useUpdateNoodlerStageProfile() {
   return useMutation({
     mutationFn: ({ accountId, ...input }: { accountId: string } & NoodleStageProfileInput) =>
       api.put<NoodlerStageProfile>(`/noodle/noodler/accounts/${encodeURIComponent(accountId)}/stage-profile`, input),
-    onSuccess: () => qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+    onSuccess: () =>
+      Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]),
   });
 }
 
@@ -134,7 +149,11 @@ export function useDeleteNoodlerStageProfile() {
       api.delete<NoodleAccount>(`/noodle/noodler/accounts/${encodeURIComponent(accountId)}`),
     onSuccess: (_account, accountId) => {
       qc.removeQueries({ queryKey: noodleKeys.privatePosts(accountId) });
-      qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() });
+      return Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateEligibleAccountsRoot() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]);
     },
   });
 }
@@ -158,11 +177,27 @@ export function useGeneratePrivateNoodlePost() {
   const qc = useQueryClient();
   return useMutation({
     mutationFn: (input: NoodlePrivateGenerationRequest) =>
-      api.post<NoodlePost>("/noodle/refresh", {
+      api.post<NoodlerManagedPost>("/noodle/refresh", {
         ...input,
         debugMode: useUIStore.getState().debugMode,
       } satisfies NoodlePrivateGenerationRequest),
-    onSuccess: (_post, input) => qc.invalidateQueries({ queryKey: noodleKeys.privatePosts(input.targetAccountId) }),
+    onSuccess: (_post, input) =>
+      Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privatePosts(input.targetAccountId) }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]),
+  });
+}
+
+export function useCreateNoodlerPost() {
+  const qc = useQueryClient();
+  return useMutation({
+    mutationFn: (input: NoodlePrivatePostCreateInput) => api.post<NoodlerManagedPost>("/noodle/noodler/posts", input),
+    onSuccess: (_post, input) =>
+      Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privatePosts(input.targetAccountId) }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]),
   });
 }
 
@@ -228,19 +263,31 @@ export function useRemoveNoodlerInteraction() {
   });
 }
 
-/** Edit a NoodleR (private stage-profile) post. Hits the private-only endpoint; the
- * caller refetches the affected viewer feed on success. */
 export function useUpdateNoodlerPost() {
+  const qc = useQueryClient();
   return useMutation({
-    mutationFn: ({ id, ...input }: { id: string } & NoodlePostUpdateInput) =>
-      api.patch<NoodlePost>(`/noodle/noodler/posts/${encodeURIComponent(id)}`, input),
+    mutationFn: ({ id, accountId: _accountId, ...input }: { id: string; accountId: string } & NoodlePrivatePostUpdateInput) =>
+      api.patch<NoodlerManagedPost>(`/noodle/noodler/posts/${encodeURIComponent(id)}`, input),
+    onSuccess: (_post, input) => {
+      return Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privatePosts(input.accountId) }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]);
+    },
   });
 }
 
-/** Delete a NoodleR (private stage-profile) post. */
 export function useDeleteNoodlerPost() {
+  const qc = useQueryClient();
   return useMutation({
-    mutationFn: (id: string) => api.delete<NoodlePost>(`/noodle/noodler/posts/${encodeURIComponent(id)}`),
+    mutationFn: ({ id }: { id: string; accountId: string }) =>
+      api.delete<NoodlerManagedPost>(`/noodle/noodler/posts/${encodeURIComponent(id)}`),
+    onSuccess: (_post, input) => {
+      return Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privatePosts(input.accountId) }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]);
+    },
   });
 }
 
@@ -260,7 +307,10 @@ export function useUpdateNoodlerAccess() {
         patch: { access },
       } satisfies NoodleAccountSettingsPatchInput),
     onSuccess: () => {
-      qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() });
+      return Promise.all([
+        qc.invalidateQueries({ queryKey: noodleKeys.privateAccounts() }),
+        qc.invalidateQueries({ queryKey: noodleKeys.privateViewers() }),
+      ]);
     },
   });
 }
